feat(phone): vehicle model name now fetches
